Add public getFormattedArea and getFormattedLength functions
