+++ exports/ol-ext/interactions/measure.js
@@ -1,9 +1,15 @@
+goog.require('ngeo.interaction.Measure');
 goog.require('ngeo.interaction.MeasureArea');
 goog.require('ngeo.interaction.MeasureAzimut');
 goog.require('ngeo.interaction.MeasureLength');
 
 
-goog.exportSymbol('ngeo.interaction.MeasureArea', ngeo.interaction.MeasureArea);
+goog.exportSymbol('ngeo.interaction.Measure.getFormattedArea',
+    ngeo.interaction.Measure.getFormattedArea);
+goog.exportSymbol('ngeo.interaction.Measure.getFormattedLength',
+    ngeo.interaction.Measure.getFormattedLength);
+goog.exportSymbol('ngeo.interaction.MeasureArea',
+    ngeo.interaction.MeasureArea);
 goog.exportSymbol('ngeo.interaction.MeasureAzimut',
     ngeo.interaction.MeasureAzimut);
 goog.exportSymbol('ngeo.interaction.MeasureLength',
