+++ src/ol-ext/interaction/measurearea.js
@@ -58,35 +58,12 @@ ngeo.interaction.MeasureArea.prototype.getDrawInteraction = function(style,
 ngeo.interaction.MeasureArea.prototype.handleMeasure = function(callback) {
   var geom = /** @type {ol.geom.Polygon} */
       (this.sketchFeature.getGeometry());
-  var output = this.formatMeasure_(geom);
+  var proj = this.getMap().getView().getProjection();
+  var output = ngeo.interaction.Measure.getFormattedArea(geom, proj);
   var verticesCount = geom.getCoordinates()[0].length;
   var coord = null;
   if (verticesCount > 2) {
     coord = geom.getInteriorPoint().getCoordinates();
   }
   callback(output, coord);
-};
-
-
-/**
- * Format measure output.
- * @param {ol.geom.Polygon} polygon
- * @return {string}
- * @private
- */
-ngeo.interaction.MeasureArea.prototype.formatMeasure_ = function(polygon) {
-  var map = this.getMap();
-  var sourceProj = map.getView().getProjection();
-  var geom = /** @type {ol.geom.Polygon} */ (polygon.clone().transform(
-      sourceProj, 'EPSG:4326'));
-  var coordinates = geom.getLinearRing(0).getCoordinates();
-  var area = Math.abs(ol.sphere.WGS84.geodesicArea(coordinates));
-  var output;
-  if (area > 1000000) {
-    output = parseFloat((area / 1000000).toPrecision(3)) +
-        ' ' + 'km<sup>2</sup>';
-  } else {
-    output = parseFloat(area.toPrecision(3)) + ' ' + 'm<sup>2</sup>';
-  }
-  return output;
 };
